+++ src/figdice/classes/ViewElementTag.php
@@ -28,7 +28,6 @@ use Psr\Log\LoggerInterface;
 use figdice\View;
 use figdice\LoggerFactory;
 use figdice\exceptions\RenderingException;
-use figdice\exceptions\DictionaryDuplicateKeyException;
 use figdice\exceptions\RequiredAttributeException;
 use figdice\exceptions\FileNotFoundException;
 
@@ -488,11 +487,6 @@ class ViewElementTag extends ViewElement implements \Serializable {
 		//fig:trans
 		if($this->name == $this->view->figNamespace . 'trans') {
 			return $this->fig_trans();
-		}
-		//================================================================
-		//fig:dictionary
-		if($this->name == $this->view->figNamespace . 'dictionary') {
-			return $this->fig_dictionary();
 		}
 
 		//================================================================
@@ -1066,80 +1060,6 @@ class ViewElementTag extends ViewElement implements \Serializable {
 		return $outputBuffer;
 	}
 
-	/**
-	 * Loads a language XML file, to be used within the current view.
-	 *
-	 * If a Temp path was specified in the View,
-	 * we try to compile (serialize) the XML key-value collection and store
-	 * the serialized form in a 'Dictionary/(langcode)' subfolder of the temp path.
-	 */
-	private function fig_dictionary() {
-		//If a @source attribute is specified,
-		//it means that when the target (view's language) is the same as @source,
-		//then don't bother loading dictionary file, nor translating: just render the tag's children.
-
-		$file = $this->attributes['file'];
-		$filename = $this->getView()->getTranslationPath() . '/' . $this->getView()->getLanguage() . '/' . $file;
-
-		$name = $this->getAttribute('name', null);
-		$source = $this->getAttribute('source', null);
-		$dictionary = new Dictionary($filename, $source);
-
-
-		if ( ($this->getView()->getLanguage() == '') || ($source == $this->getView()->getLanguage()) ) {
-			// If the current View does not specify a Language,
-			// or if the dictionary to load is same language as View,
-			// let's not care about i18n.
-			// We will activate i18n only if the dictionary explicitly specifies a source,
-			// which means that we cannot simply rely on contents of the fig:trans tags.
-			// However, we still need to hook the Dictionary object as a placeholder,
-			// so that subsequent trans tag for the given dic name and source will
-			// simply render their contents.
-			$this->getCurrentFile()->addDictionary($dictionary, $name);
-			return '';
-		}
-
-		//TODO: Please optimize here: cache the realpath of the loaded dictionaries,
-		//so as not to re-load an already loaded dictionary in same View hierarchy.
-
-
-		try {
-			//Determine whether this dictionary was pre-compiled:
-			if($this->getView()->getTempPath()) {
-				$tmpFile = $this->getView()->getTempPath() . '/' . 'Dictionary' . '/' . $this->getView()->getLanguage() . '/' . $file . '.php';
-				//If the tmp file already exists,
-				if(file_exists($tmpFile)) {
-					//but is older than the source file,
-					if(filemtime($tmpFile) < filemtime($filename)) {
-						Dictionary::compile($filename, $tmpFile);
-					}
-				}
-				else {
-					Dictionary::compile($filename, $tmpFile);
-				}
-				$dictionary->restore($tmpFile);
-			}
-
-			//If we don't even have a temp folder specified, load the dictionary for the first time.
-			else {
-					$dictionary->load();
-			}
-		} catch(FileNotFoundException $ex) {
-			throw new FileNotFoundException('Translation file not found: file=' . $filename . 
-				', language=' . $this->getView()->getLanguage() . 
-				', source=' . $this->getCurrentFilename(), 
-				$this->getCurrentFilename() );
-		} catch(DictionaryDuplicateKeyException $ddkex) {
-			$this->getLogger()->error('Duplicate key: "' . $ddkex->getKey() . '" in dictionary: ' . $ddkex->getFilename());
-		}
-
-
-		//Hook the dictionary to the current file.
-		//(in fact this will bubble up the message as high as possible, ie:
-		//to the highest parent which does not bear a dictionary of same name)
-		$this->getCurrentFile()->addDictionary($dictionary, $name);
-		return '';
-	}
 
 	/**
 	 * Translates a caption given its key and dictionary name.
@@ -1312,7 +1232,7 @@ class ViewElementTag extends ViewElement implements \Serializable {
 	 *
 	 * @return LoggerInterface
 	 */
-	private function getLogger() {
+	protected function getLogger() {
 		if(! $this->logger) {
 			$this->logger = LoggerFactory::getLogger(get_class($this));
 		}
